Create an account for an organization
In the multi-tenancy chapter, it says the following:
> If at some future point you want to disable that tenant, all you have to do is remove the Account Store Mapping between that Organization and your Application, and the tenant’s users would no longer be able to log in.

This will only work if there is no direct mapping between application and directory, but only through `application-->organisation-->(tenant)group-->directory`. To make this feasible, one would expect to be able to create accounts directly for an organisation, as opposed to a group or directory. To my surprise, this is actually possible, although it's not documented. This PR changes that.

I'd propose to also clarify this in the multi-tenancy chapter.